+++ src/themes/teams/components/Segment/segmentStyles.ts
@@ -1,13 +1,21 @@
-import { pxToRem } from '../../../../lib'
-import { ICSSInJSStyle } from '../../../types'
+import { ICSSInJSStyle, ComponentSlotStylesInput } from '../../../types'
+import { SegmentProps } from 'src/components/Segment/Segment'
+import { SegmentVariables } from './segmentVariables'
 
-export default {
-  root: (): ICSSInJSStyle => {
+const segmentStyles: ComponentSlotStylesInput<SegmentProps, SegmentVariables> = {
+  root: ({ props: { inverted }, variables: v }): ICSSInJSStyle => {
     return {
-      padding: '1em',
-      boxShadow: '0 1px 2px 0 rgba(34,36,38,.15)',
-      border: '1px solid rgba(34,36,38,.15)',
-      borderRadius: pxToRem(5),
+      padding: '1rem',
+      boxShadow: '0 1px 1px 1px rgba(34,36,38,.15)',
+      ...(inverted
+        ? {
+            backgroundColor: v.color,
+          }
+        : v.color && {
+            borderTop: `2px solid ${v.color}`,
+          }),
     }
   },
 }
+
+export default segmentStyles