+++ src/themes/teams/components/Segment/segmentVariables.ts
@@ -1,0 +1,5 @@
+export interface SegmentVariables {
+  color: string
+}
+
+export default () => ({ color: undefined })